Add `run-server` function
The run-server function is a way to create an runnable app and run it outside of
the default mount state. With this function, users can define their own states
on their apps and benefit from automatic reload in development.

Mount has a very interesting feature which reloads the state everytime the
defstate form gets evaluated at the repl, but when you hide the defstate in a
third party library, you cannot benefit from that, as you don't have direct
access to it's source code.

This is an attempt to make rama more flexible, not requesting too much changes
in existing apps to use it with the version 1.1.